Split up embedded classes into their own .cs files
